+++ spec/Select.spec.js
@@ -3,21 +3,38 @@ describe("SELECT", function () {
 
     it("✔️ constructed", function () {
       var container = new shiva.Select({
-        options: ["hello", "goodbye"]
+        options: [
+          {value:"hello", text:"shiva"}, 
+          {value:"goodbye", text:"shiva"}
+        ]
       });
-
-      expect(container.value).toEqual("hello");
+      
+      var option = container.element.children.item(0);
+      expect(option.innerHTML).toEqual("shiva");
     });
 
     it("✔️ get selected index", function () {
       var container = new shiva.Select({
-        options: ["hello", "goodbye"]
+        options: [
+          {value:"hello", text:"shiva"}, 
+          {value:"goodbye", text:"shiva"}
+        ]
       });
 
       expect(container.selectedIndex).toEqual(0);
     });
 
-
+    it("✔️ set selected index", function () {
+      var container = new shiva.Select({
+        options: [
+          {value:"hello", text:"shiva"}, 
+          {value:"goodbye", text:"shiva"}
+        ]
+      });
+      container.selectedIndex = 1;
+ 
+      expect(container.selectedIndex).toEqual(1);
+    });
 
   });
 });